Fix setMatrix block

+++ src/index.js
@@ -663,7 +663,7 @@ import BetterQuakeIcon from './assets/BetterQuakeIcon.svg'
 
       if (!Array.isArray(converted)) return
       converted = converted.map(function (str) {
-        return parseInt(str)
+        return parseFloat(str)
       })
 
       drawable.BetterQuake.uniforms[UNIFORM] = converted
